Move PipelineConfig class to beam package
